chore(secrets): separate ffi logic in macOS code
