+++ packages/secrets/src/keyring/src/os/mac/mod.rs
@@ -2,12 +2,17 @@ use super::error::KeyringError;
 
 mod keychain;
 mod keychain_item;
+mod ffi;
+mod error;
+
+use error::Error;
 
 use keychain::Keychain;
+
 const ERR_SEC_ITEM_NOT_FOUND: i32 = -25300;
 
-impl From<keychain::Error> for KeyringError {
-    fn from(error: keychain::Error) -> Self {
+impl From<Error> for KeyringError {
+    fn from(error: Error) -> Self {
         KeyringError::Library {
             name: "security_framework".to_owned(),
             details: format!("{:?}", error.message()),
@@ -15,12 +20,12 @@ impl From<keychain::Error> for KeyringError {
     }
 }
 
-/// 
+///
 /// Attempts to set a password for a given service and account.
-/// 
+///
 /// - `service`: The service name for the new credential
 /// - `account`: The account name for the new credential
-/// 
+///
 /// Returns:
 /// - `true` if the credential was stored successfully
 /// - A `KeyringError` if there were any issues interacting with the credential vault
@@ -37,16 +42,16 @@ pub fn set_password(
     }
 }
 
-/// 
+///
 /// Returns a password contained in the given service and account, if found.
-/// 
+///
 /// - `service`: The service name that matches the credential of interest
 /// - `account`: The account name that matches the credential of interest
-/// 
+///
 /// Returns:
 /// - `Some(password)` if a matching credential was found; `None` otherwise
 /// - A `KeyringError` if there were any issues interacting with the credential vault
-/// 
+///
 pub fn get_password(service: &String, account: &String) -> Result<Option<String>, KeyringError> {
     let keychain = Keychain::default().unwrap();
     match keychain.find_password(service.as_str(), account.as_str()) {
@@ -56,15 +61,15 @@ pub fn get_password(service: &String, account: &String) -> Result<Option<String>
     }
 }
 
-/// 
+///
 /// Returns the first password (if any) that matches the given service pattern.
-/// 
+///
 /// - `service`: The service pattern that matches the credential of interest
-/// 
+///
 /// Returns:
 /// - `Some(password)` if a matching credential was found; `None` otherwise
 /// - A `KeyringError` if there were any issues interacting with the credential vault
-/// 
+///
 pub fn find_password(service: &String) -> Result<Option<String>, KeyringError> {
     let cred_attrs: Vec<&str> = service.split("/").collect();
     if cred_attrs.len() < 2 {
@@ -85,16 +90,16 @@ pub fn find_password(service: &String) -> Result<Option<String>, KeyringError> {
     }
 }
 
-/// 
+///
 /// Attempts to delete the password associated with a given service and account.
-/// 
+///
 /// - `service`: The service name of the credential to delete
 /// - `account`: The account name of the credential to delete
-/// 
+///
 /// Returns:
 /// - `true` if a matching credential was deleted; `false` otherwise
 /// - A `KeyringError` if there were any issues interacting with the credential vault
-/// 
+///
 pub fn delete_password(service: &String, account: &String) -> Result<bool, KeyringError> {
     let keychain = Keychain::default().unwrap();
     match keychain.find_password(service.as_str(), account.as_str()) {
@@ -107,16 +112,16 @@ pub fn delete_password(service: &String, account: &String) -> Result<bool, Keyri
     }
 }
 
-/// 
+///
 /// Builds a vector of all credentials matching the given service pattern.
-/// 
+///
 /// - `service`: The service pattern that matches the credential(s) of interest
 /// - `credentials`: The vector consisting of (username, password) pairs for each credential that matches 
-/// 
+///
 /// Returns:
 /// - `true` if at least 1 credential was found, `false` otherwise
 /// - A `KeyringError` if there were any issues interacting with the credential vault
-/// 
+///
 pub fn find_credentials(
     service: &String,
     credentials: &mut Vec<(String, String)>,
